+++ http4k-contract/src/main/kotlin/org/http4k/contract/OpenApi.kt
@@ -7,7 +7,7 @@ import org.http4k.core.with
 import org.http4k.format.Json
 import org.http4k.format.JsonErrorResponseRenderer
 import org.http4k.lens.Failure
-import org.http4k.lens.Header.Common.CONTENT_TYPE
+import org.http4k.lens.Header.CONTENT_TYPE
 import org.http4k.lens.Meta
 import org.http4k.util.JsonSchema
 import org.http4k.util.JsonToJsonSchema