+++ http4k-core/src/main/kotlin/org/http4k/lens/header.kt
@@ -12,18 +12,24 @@ object Header : BiDiLensSpec<HttpMessage, String>("header", StringParam,
     LensGet { name, target -> target.headerValues(name).map { it ?: "" } },
     LensSet { name, values, target -> values.fold(target) { m, next -> m.header(name, next) } }
 ) {
-    object Common {
-        val CONTENT_TYPE = map(
-            {
-                val parts = it.split(";")
-                if (parts.size == 2) {
-                    val directive = parts[1].split("=")
-                    if (directive.size == 2) ContentType(parts[0].trim(), directive[0].trim() to directive[1].trim())
-                    else ContentType(parts[0].trim())
-                } else ContentType(it.trim())
-            },
-            ContentType::toHeaderValue).optional("content-type")
+    val CONTENT_TYPE = map(
+        {
+            val parts = it.split(";")
+            if (parts.size == 2) {
+                val directive = parts[1].split("=")
+                if (directive.size == 2) ContentType(parts[0].trim(), directive[0].trim() to directive[1].trim())
+                else ContentType(parts[0].trim())
+            } else ContentType(it.trim())
+        },
+        ContentType::toHeaderValue).optional("content-type")
 
-        val LOCATION = map({ of(it) }, Uri::toString).required("location")
+    val LOCATION = map({ of(it) }, Uri::toString).required("location")
+
+    object Common {
+        @Deprecated("moved to main Header object", ReplaceWith("Header.CONTENT_TYPE"))
+        val CONTENT_TYPE = Header.CONTENT_TYPE
+
+        @Deprecated("moved to main Header object", ReplaceWith("Header.LOCATION"))
+        val LOCATION = Header.LOCATION
     }
 }