+++ http4k-core/src/test/kotlin/org/http4k/lens/BodyTest.kt
@@ -9,7 +9,7 @@ import org.http4k.core.ContentType.Companion.TEXT_PLAIN
 import org.http4k.core.Method
 import org.http4k.core.Request
 import org.http4k.core.with
-import org.http4k.lens.Header.Common.CONTENT_TYPE
+import org.http4k.lens.Header.CONTENT_TYPE
 import org.junit.jupiter.api.Test
 
 class BodyTest {