+++ http4k-k8s/src/main/kotlin/org/http4k/k8s/Health.kt
@@ -6,7 +6,7 @@ import org.http4k.core.Response
 import org.http4k.core.Status.Companion.OK
 import org.http4k.core.Status.Companion.SERVICE_UNAVAILABLE
 import org.http4k.core.with
-import org.http4k.lens.Header.Common.CONTENT_TYPE
+import org.http4k.lens.Header.CONTENT_TYPE
 import org.http4k.routing.bind
 import org.http4k.routing.routes