+++ http4k-multipart/src/test/kotlin/org/http4k/core/MultipartIteratorTest.kt
@@ -3,7 +3,8 @@ package org.http4k.core
 import com.natpryce.hamkrest.assertion.assertThat
 import com.natpryce.hamkrest.equalTo
 import com.natpryce.hamkrest.throws
-import org.http4k.lens.Header
+import org.http4k.core.ContentType.Companion.TEXT_PLAIN
+import org.http4k.lens.Header.CONTENT_TYPE
 import org.http4k.multipart.AlreadyClosedException
 import org.junit.jupiter.api.Test
 import java.io.InputStream
@@ -14,11 +15,11 @@ class MultipartIteratorTest {
     @Test
     fun `can stream multiparts`() {
         val form = MultipartFormBody("bob") + ("field" to "bar") +
-            ("file" to FormFile("foo.txt", ContentType.TEXT_PLAIN, "content".byteInputStream())) +
+            ("file" to FormFile("foo.txt", TEXT_PLAIN, "content".byteInputStream())) +
             ("field2" to "bar2")
 
         val req = Request(Method.POST, "")
-            .with(Header.Common.CONTENT_TYPE of ContentType.MultipartFormWithBoundary(form.boundary))
+            .with(CONTENT_TYPE of ContentType.MultipartFormWithBoundary(form.boundary))
             .body(form)
 
         val fileStream = AtomicReference<InputStream>()